nxp-s32g: update to support HSE feature based on NXP SDK bsp32
Issue: LINUXEXEC-19572

Given the HSE enable steps have been changed, the bsp layer should be updated
to compatible with it, and this patch is to do this work.

In detail, the updates as following:

  1. add hse.inc file to define the related variables needed for HSE
  2. update uboot recipe:
     a. add a new patch to set CONFIG_HSE_SECBOOT for s32g2xxaevb when
        HSE enabled
     b. correct SECBOOT to the location of HSE firmware in u-boot-s32.cfgout
     c. drop the unneeded do_compile_append()
  3. update atf recipe:
     a. add "FIP_MEMORY_OFFSET=0x3407e910 HSE_SECBOOT=1" compile parameters
        for s32g2xxaevb when HSE enabled to update binary layout for HSE
     b. add RSA keys generation and RSA sign operations of atf binaries to
        compatible with new check steps for HSE
  4. add related RSA public key into boot partition for HSE verification
  5. add new dd operations for s32g2xxaevb when HSE enabled in file
     make_nxp_s32g_wic.inc

And now the HSE support is only avialable for s32g2xxaevb, and there will be
updates for s32g274ardb2 and s32g399ardb3 in future. Moreover, the README
files are updated too.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>